==> house_price_pred/notebooks/02-sjh-latitude-longitude.ipynb <==
# visualization
import matplotlib.pyplot as plt
import matplotlib.font_manager as fm
fe = fm.FontEntry(
    fname=r'/usr/share/fonts/truetype/nanum/NanumGothic.ttf', # ttf 파일이 저장되어 있는 경로
    name='NanumBarunGothic')                        # 이 폰트의 원하는 이름 설정
fm.fontManager.ttflist.insert(0, fe)              # Matplotlib에 폰트 추가
plt.rcParams.update({'font.size': 10, 'font.family': 'NanumBarunGothic'}) # 폰트 설정
plt.rc('font', family='NanumBarunGothic')
import seaborn as sns

# utils
import pandas as pd
import numpy as np
from tqdm import tqdm
import pickle
import warnings;warnings.filterwarnings('ignore')

# Model
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error
from sklearn.ensemble import RandomForestRegressor
from sklearn import metrics

import eli5
from eli5.sklearn import PermutationImportance

pd.set_option('display.max_columns', None)

train = pd.read_csv("/data/ephemeral/home/train.csv")

def get_gu(addr):
    idx = addr.find('구')
    return addr[:idx+1]


train['시구'] = train['시군구'].apply(get_gu)

# 서울특별시 강남구 테해란로 3 형식으로 주소 변환 -> geocoder에 넣기 위함임
train['addr'] = train['시구'] + ' ' + train['도로명']

%pip install geopy

# import 라이브러리
from geopy.geocoders import Nominatim

def geocoding(address):
    try:
        geo_local = Nominatim(user_agent='South Korea')  #지역설정
        location = geo_local.geocode(address)
        geo = [location.latitude, location.longitude]
        return geo

    except:
        return [0,0]

# 시작과 끝 지점 입력
start = 
end = 

train_point = train.loc[(train['좌표X'].isna()) | (train['좌표Y'].isna()), :]
train_point = train_point.reset_index()
train_point = train_point.iloc[start:end]
len(train_point)

# 각자 이름 써주세요
name = ''

# x -> 경도 -> longitude
# y -> 위도 -> latitude

for idx,addr in enumerate(tqdm(train_point.addr)):
    point = geocoding(addr)
    train_point.loc[idx,'좌표Y'] = point[0]
    train_point.loc[idx,'좌표X'] = point[1]
    if idx / 1000 == 0:
        train_point.to_csv(f"/data/ephemeral/home/train_xy_{name}.csv", index=False)    # 하다가 터질 수 있으니깐 백업

# 최종 결과물 저장
train_point.to_csv(f"/data/ephemeral/home/train_xy_{name}.csv", index=False)

#### 필독 ####

'''
geocoder API 오류로 인해 return 값이 [0, 0]이 되어 좌표X, Y가 0으로 될 수도 있음

이러한 경우를 꼭 확인합시다 !!

'''
