==> house_price_pred/notebooks/05-lyh-subway.ipynb <==
# visualization
import matplotlib.pyplot as plt
import matplotlib.font_manager as fm
fe = fm.FontEntry(
    fname=r'/usr/share/fonts/truetype/nanum/NanumGothic.ttf', # ttf 파일이 저장되어 있는 경로
    name='NanumBarunGothic')                        # 이 폰트의 원하는 이름 설정
fm.fontManager.ttflist.insert(0, fe)              # Matplotlib에 폰트 추가
plt.rcParams.update({'font.size': 10, 'font.family': 'NanumBarunGothic'}) # 폰트 설정
plt.rc('font', family='NanumBarunGothic')
import seaborn as sns

# utils
import pandas as pd
import numpy as np
from tqdm import tqdm
import pickle
import warnings;warnings.filterwarnings('ignore')

# Model
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error
from sklearn.ensemble import RandomForestRegressor
from sklearn import metrics

import eli5
from eli5.sklearn import PermutationImportance

import random

train_path = '../data/interim/train_final_ver1.csv'
test_path = '../data/interim/test_final_ver1.csv'

df_train = pd.read_csv(train_path, index_col=0)
df_test = pd.read_csv(test_path, index_col=0)

subway_path = '../data/external/subway_feature.csv'

df_subway = pd.read_csv(subway_path)

display(df_subway)

df_subway = df_subway.rename(columns={'위도' : 'sub_y', '경도' : 'sub_x'})
display(df_subway)

df_train['is_test'] = 0
df_test['is_test'] = 1
df_all = pd.concat([df_train, df_test])
display(df_all)

display(df_all.describe())
display(df_subway.describe())

df_all.loc[(df_all['아파트명'] == '삼성') & (df_all['x'] < 1), 'x'] = 126.9788
df_all.loc[(df_all['아파트명'] == '삼성') & (df_all['y'] < 1), 'y'] = 37.6133

display(df_all.describe())
display(df_subway.describe())

import math

# 위경도를 이용해 두 지점간의 거리를 구하는 함수를 생성합니다.
def haversine_distance(lat1, lon1, lat2, lon2): # y1, x1, y2, x2
    radius = 6371.0

    lat1 = math.radians(lat1)
    lon1 = math.radians(lon1)
    lat2 = math.radians(lat2)
    lon2 = math.radians(lon2)

    dlon = lon2 - lon1
    dlat = lat2 - lat1
    a = math.sin(dlat / 2) ** 2 + math.cos(lat1) * math.cos(lat2) * math.sin(dlon / 2) ** 2
    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1 - a))

    distance = radius * c
    return distance

df_loc = df_all[['x', 'y']]
unique_loc = set()
for i in range(len(df_loc)):
    unique_loc.add((df_loc['x'].values[i], df_loc['y'].values[i]))

unique_list = list(unique_loc)

unique_x, unique_y = [], []
for i in range(len(unique_list)):
    x, y = unique_list[i]
    unique_x.append(x)
    unique_y.append(y)

ser_x = pd.Series(unique_x)
ser_y = pd.Series(unique_y)
display(ser_x)
display(ser_y)

df_loc = pd.DataFrame({'x' : ser_x, 'y' : ser_y})

display(df_loc)
display(df_subway)

subway_sample = df_subway.iloc[:10]
loc_sample = df_loc.iloc[:10]

display(subway_sample)
display(loc_sample)

def get_subway_min_dist(df1, df2): # 거리 차이를 계산할 데이터 프레임 --> 전체 데이터, 버스 or 지하철 데이터
    min_dist = []
    for i in range(len(df1)):
        x, y = df1['x'].values[i], df1['y'].values[i]
        dist = []
        for j in range(len(df2)):
            tar_x, tar_y = df2['sub_x'].values[j], df2['sub_y'].values[j]
            harv_dist = haversine_distance(y, x, tar_y, tar_x)
            dist.append(harv_dist)
        min_dist.append(min(dist))

    return min_dist
        

# sample로 결과값이 잘 나오는지  테스트
sub_min_dist = get_subway_min_dist(loc_sample, subway_sample)
print(sub_min_dist)

# 전체 데이터셋에 대해 실행
sub_min_dist = get_subway_min_dist(df_loc, df_subway)

df_loc['sub_min_dist'] = sub_min_dist
display(df_loc)

df_subway = pd.merge(df_all, df_loc, on=['x', 'y'], how='inner')
display(df_subway)

# 결측치 확인
df_subway.isnull().sum()

subway_train = df_subway[df_subway['is_test'] == 0]
subway_test = df_subway[df_subway['is_test'] == 1]

display(subway_train)
display(subway_test)

subway_train.to_csv('../data/interim/subway_train.csv', index=False)
subway_test.to_csv('../data/interim/subway_test.csv', index=False)
