==> house_price_pred/notebooks/06-lyh-merge-final_bus_subway.ipynb <==
import pandas as pd

train_path = '../data/interim/merge_train_final.csv'
test_path = '../data/interim//merge_test_final.csv'

bus_train_path = '../data/interim/bus_train.csv'
bus_test_path = '../data/interim/bus_test.csv'

subway_train_path = '../data/interim/subway_train.csv'
subway_test_path = '../data/interim/subway_test.csv'

df_train = pd.read_csv(train_path, index_col=0)
df_test = pd.read_csv(test_path, index_col=0)

bus_train = pd.read_csv(bus_train_path)
bus_test = pd.read_csv(bus_test_path)

subway_train = pd.read_csv(subway_train_path)
subway_test = pd.read_csv(subway_test_path)

display(df_train)
display(df_test)
print('-' * 50)
display(bus_train)
display(bus_test)
print('-' * 50)
display(subway_train)
display(subway_test)

df_train['is_test'] = 0
df_test['is_test'] = 1
df_all = pd.concat([df_train, df_test])
bus_all = pd.concat([bus_train, bus_test])
subway_all = pd.concat([subway_train, subway_test])

display(df_all.head(1))
display(bus_all.head(1))
display(subway_all.head(1))

display(df_all.shape)
display(bus_all.shape)
display(subway_all.shape)

df_all['bus_min_dist'] = bus_all['bus_min_dist']
df_all['sub_min_dist'] = subway_all['sub_min_dist']
display(df_all)

df_all_train = df_all[df_all['is_test'] == 0]
df_all_test = df_all[df_all['is_test'] == 1]

display(df_all_train)
display(df_all_test)

del df_all_train['is_test']
del df_all_test['is_test']

df_all_train.to_csv('../data/interim/final_bus_subway_train.csv', index=False)
df_all_test.to_csv('../data/interim/final_bus_subway_test.csv', index=False)
