==> house_price_pred/notebooks/14-SJY_h2o.ipynb <==
!pip install h2o
# 한글 폰트 사용을 위한 라이브러리입니다.
!apt-get install -y fonts-nanum

# visualization
import matplotlib.pyplot as plt
import matplotlib.font_manager as fm
fe = fm.FontEntry(
    fname=r'/usr/share/fonts/truetype/nanum/NanumGothic.ttf', # ttf 파일이 저장되어 있는 경로
    name='NanumBarunGothic')                        # 이 폰트의 원하는 이름 설정
fm.fontManager.ttflist.insert(0, fe)              # Matplotlib에 폰트 추가
plt.rcParams.update({'font.size': 10, 'font.family': 'NanumBarunGothic'}) # 폰트 설정
plt.rc('font', family='NanumBarunGothic')
import seaborn as sns

# utils
import pandas as pd
import numpy as np
from tqdm import tqdm
import pickle
import warnings;warnings.filterwarnings('ignore')

# Model
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error
from sklearn.ensemble import RandomForestRegressor

import h2o
from h2o.automl import H2OAutoML

# 필요한 데이터를 load
total_path = "../total-data_ver2.csv"

dt_total = pd.read_csv(total_path)

# is_test 컬럼 값이 0인 행을 train 데이터셋으로 저장
df_train = dt_total[dt_total['is_test'] == 0]

# is_test 컬럼 값이 1인 행을 test 데이터셋으로 저장
df_test = dt_total[dt_total['is_test'] == 1]

# '계약년월'에서 '연도' 추출 후 int 형으로 변환하여 새로운 열 추가
df_train = df_train.assign(year=df_train['계약년월'].astype('str').str[:4].astype(int)).query('year > 2020')

# '계약년월'에서 '연도' 추출 후 int 형으로 변환하여 새로운 열 추가
df_test = df_test.assign(year=df_test['계약년월'].astype('str').str[:4].astype(int))

# '계약년월'에서 '월' 추출 후 5로 나눈 값을 새로운 열로 추가
df_train, df_test = map(lambda df: df.assign(계약월=lambda x: (x['계약년월'] % 100) // 5), [df_train, df_test])

# 타겟 변수 생성
df_train['price'] = df_train.groupby(['아파트명', '계약월'])['target'].transform('mean')

# 중복 데이터 제거
df_train = df_train.drop_duplicates(subset=['아파트명', '계약월'])

import pandas as pd

# 범주형 변수를 원-핫 인코딩으로 변환
df_train_encoded = pd.get_dummies(df_train)
df_test_encoded = pd.get_dummies(df_test)

# 학습 데이터와 테스트 데이터에 있는 모든 변수가 동일하도록 조정
df_train_encoded, df_test_encoded = df_train_encoded.align(df_test_encoded, join='left', axis=1)

# 결측값을 0으로 채움 (학습 데이터에 없는 범주형 레벨 때문에 발생)
df_test_encoded.fillna(0, inplace=True)

# H2O 서버 초기화
h2o.init()

# 원-핫 인코딩된 데이터 프레임을 H2O용 프레임으로 변환
h2o_train = h2o.H2OFrame(df_train_encoded)
h2o_test = h2o.H2OFrame(df_test_encoded)

# 타겟 변수와 입력 변수 지정
x = ['전용면적', 'x'] # 사용하고자 하는 변수만 지정
y = "price"

# AutoML 객체 생성
aml = H2OAutoML(max_models = 10, seed = 1)

# 모델 학습
aml.train(x = x, y = y, training_frame = h2o_train)

# 테스트 데이터에 대한 예측
preds = aml.predict(h2o_test)

# print(pred.columns)

lb = aml.leaderboard
lb.head(rows=lb.nrows)

lb = h2o.automl.get_leaderboard(aml, extra_columns = 'ALL')

lb

# 예측 결과를 pandas 데이터프레임으로 변환
preds_df = preds.as_data_frame()

# 데이터프레임을 CSV 파일로 저장
preds_df.to_csv('predictions.csv', index=False)
